+++ script/source/classes/MouseMove.js
@@ -112,8 +112,8 @@ class MouseMove
             Author:    'Justin Don Byrne',
             Created:   'Aug, 04 2023',
             Library:   'Mouse Move: Automated mouse cursor for web presentation',
-            Updated:   'Sep, 06 2023',
-            Version:   '0.1.8',
+            Updated:   'Sep, 07 2023',
+            Version:   '0.1.9',
             Copyright: 'Copyright (c) 2023 Justin Don Byrne'
         }
     }